+++ packages/uploader/lib/utils/index.ts
@@ -1,9 +1,0 @@
-export const clearBlank = () => {}
-
-export const getRequestParams = () => {
-  
-}
-
-export const createSignature = ({ params, apiSecret }: { params: string[], apiSecret: string }) => {
-
-}