+++ packages/url/lib/transformers/position.ts
@@ -1,4 +1,4 @@
-import { Position } from '@cld-apis/types'
+import type { Position } from '@cld-apis/types'
 import { formatValue } from './expression'
 
 export const position = ({ x, y }: Position) => {