+++ packages/utils/lib/constants/delivery.ts
@@ -7,7 +7,12 @@ export const STORAGE_TYPES = {
   FACEBOOK: 'facebook',
   TWITTER: 'twitter',
   YOUTUBE: 'youtube',
-  VIMEO: 'vimeo'
+  VIMEO: 'vimeo',
+  MULTI: 'multi',
+  INSTAGRAM: 'instagram',
+  HULU: 'hulu',
+  ANIMOTO: 'animoto',
+  DAILYMOTION: 'dailymotion'
 } as const
 
 export const RESOURCE_TYPES = {
@@ -19,5 +24,37 @@ export const RESOURCE_TYPES = {
 export const VIDEO_SOURCE_TYPES = {
   WEBM: 'webm', 
   MP4: 'mp4',
-  OGV: 'ogv'
+  OGV: 'ogv',
+  FLV: 'flv',
+  M3U8: 'm3u8',
+  TS: 'ts',
+  MOV: 'mov',
+  MKV: 'mkv',
+  MPD: 'mpd'
 } as const
+
+export const ImageFormats = {
+  GIF: "gif",
+  PNG: "png",
+  JPG: "jpg",
+  BMP: "bmp",
+  ICO: "ico",
+  PDF: "pdf",
+  TIFF: "tiff",
+  EPS: "eps",
+  JPC: "jpc",
+  JP2: "jp2",
+  PSD: "psd",
+  WEBP: "webp",
+  ZIP: "zip",
+  SVG: "svg",
+  WDP: "wdp",
+  HDX: "hpx",
+  DJVU: "djvu",
+  AI: "ai",
+  FLIF: "flif",
+  BPG: "bpg",
+  MIFF: "miff",
+  TGA: "tga",
+  HEIC: "heic",
+}