+++ packages/url/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/mayashavin/cloudinary-api/compare/cloudinary-build-url@0.2.0...cloudinary-build-url@0.2.1) (2021-02-03)
+
+
+### Bug Fixes
+
+* repeat import Asset ([2f7a1d6](https://github.com/mayashavin/cloudinary-api/commit/2f7a1d64782b8086635b28b6423c15bae332be42))
+
+
+
+
+
+# [0.2.0](https://github.com/mayashavin/cloudinary-api/compare/cloudinary-build-url@0.0.3...cloudinary-build-url@0.2.0) (2021-02-03)
+
+
+### Bug Fixes
+
+* add colorify and artistic filters ([c2409ab](https://github.com/mayashavin/cloudinary-api/commit/c2409abd302388307813b1bfe79843d1c081fa1c))
+* improve build ([a6aaf53](https://github.com/mayashavin/cloudinary-api/commit/a6aaf534e906da83b56d0c02ca0ae09be1f1f9e3))
+
+
+
+## 0.1.1 (2020-12-15)
+
+
+### Bug Fixes
+
+* logic for modifeir ([eef1514](https://github.com/mayashavin/cloudinary-api/commit/eef1514209acfbd75b9b0586cc8465890a96a706))
+* move [@cld-utils](https://github.com/cld-utils) to dependencies ([76d2e7b](https://github.com/mayashavin/cloudinary-api/commit/76d2e7bc41526c8b6707c6e9e322dd01b8f6ad8d))
+* remove socialCard ([e16ec2e](https://github.com/mayashavin/cloudinary-api/commit/e16ec2efb339dfa08c92eebf2ae2fee8e32b4b9f))
+* rename consts ([ea5a424](https://github.com/mayashavin/cloudinary-api/commit/ea5a424d9859add8e2fc9df073c0c54bf9ecc289))
+
+
+### Features
+
+* expose transformer api, add types ([049f3f2](https://github.com/mayashavin/cloudinary-api/commit/049f3f2c76fe36368fb50f689a4233f02c8122aa))
+* group x,y to position ([19f5c50](https://github.com/mayashavin/cloudinary-api/commit/19f5c5099d882604eba313077189b43d377a8a76))
+
+
+
+
+
 ## [0.0.3](https://github.com/mayashavin/cloudinary-api/compare/cloudinary-build-url@0.0.2...cloudinary-build-url@0.0.3) (2020-11-15)